+++ app/Http/Controllers/Admin/RoleController.php
@@ -6,9 +6,13 @@ use App\Role; // Perfil
 use App\Permission; // Permissões
 use App\Perpage;
 
+use Response;
+
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Session;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\DB;
 
 class RoleController extends Controller
 {
@@ -33,6 +37,10 @@ class RoleController extends Controller
      */
     public function index()
     {
+        if (Gate::denies('role.index')) {
+            abort(403, 'Acesso negado.');
+        }
+
         $roles = new Role;
 
         // filtros
@@ -72,6 +80,10 @@ class RoleController extends Controller
      */
     public function create()
     {
+        if (Gate::denies('role.create')) {
+            abort(403, 'Acesso negado.');
+        }
+
         // listagem de perfis (roles)
         $permissions = Permission::orderBy('name','asc')->get();
 
@@ -116,6 +128,10 @@ class RoleController extends Controller
      */
     public function show($id)
     {
+        if (Gate::denies('role.show')) {
+            abort(403, 'Acesso negado.');
+        }
+
         // perfil que será exibido e pode ser excluido
         $role = Role::findOrFail($id);
 
@@ -130,6 +146,10 @@ class RoleController extends Controller
      */
     public function edit($id)
     {
+        if (Gate::denies('role.edit')) {
+            abort(403, 'Acesso negado.');
+        }
+
         // perfil que será alterado
         $role = Role::findOrFail($id);
 
@@ -188,10 +208,73 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
+        if (Gate::denies('role.delete')) {
+            abort(403, 'Acesso negado.');
+        }
+
         Role::findOrFail($id)->delete();
 
         Session::flash('deleted_role', 'Permissão excluída com sucesso!');
 
         return redirect(route('roles.index'));
+    }
+
+    /**
+     * Exportação para planilha (csv)
+     *
+     * @param  int  $id
+     * @return Response::stream()
+     */
+    public function exportcsv()
+    {
+        if (Gate::denies('role.export')) {
+            abort(403, 'Acesso negado.');
+        }
+
+        $headers = [
+                'Cache-Control'       => 'must-revalidate, post-check=0, pre-check=0'
+            ,   'Content-type'        => 'text/csv'
+            ,   'Content-Disposition' => 'attachment; filename=Perfis_' .  date("Y-m-d H:i:s") . '.csv'
+            ,   'Expires'             => '0'
+            ,   'Pragma'              => 'public'
+        ];
+
+        $roles = DB::table('roles');
+
+        $roles = $roles->select('name', 'description');
+
+        // filtros
+        if (request()->has('name')){
+            $roles = $roles->where('name', 'like', '%' . request('name') . '%');
+        }
+
+        if (request()->has('description')){
+            $roles = $roles->where('description', 'like', '%' . request('description') . '%');
+        }
+
+        $roles = $roles->orderBy('name', 'asc');
+
+        $list = $roles->get()->toArray();
+
+        // nota: mostra consulta gerada pelo elloquent
+        // dd($distritos->toSql());
+
+        # converte os objetos para uma array
+        $list = json_decode(json_encode($list), true);
+
+        # add headers for each column in the CSV download
+        array_unshift($list, array_keys($list[0]));
+
+       $callback = function() use ($list)
+        {
+            $FH = fopen('php://output', 'w');
+            fputs($FH, $bom = ( chr(0xEF) . chr(0xBB) . chr(0xBF) ));
+            foreach ($list as $row) {
+                fputcsv($FH, $row, chr(9));
+            }
+            fclose($FH);
+        };
+
+        return Response::stream($callback, 200, $headers);
     }
 }
